refactor HardenXStreamCodemodTestProjectProvider
