+++ core-codemods/src/test/java/io/codemodder/codemods/HardenXStreamCodemodTest.java
@@ -1,7 +1,6 @@
 package io.codemodder.codemods;
 
 import io.codemodder.testutils.CodemodTestMixin;
-import io.codemodder.testutils.HardenXStreamCodemodTestProjectProvider;
 import io.codemodder.testutils.Metadata;
 
 @Metadata(
@@ -14,5 +13,20 @@ final class HardenXStreamCodemodTest implements CodemodTestMixin {}
     codemodType = HardenXStreamCodemod.class,
     testResourceDir = "harden-xstream-denyTypes",
     dependencies = "io.github.pixee:java-security-toolkit:1.1.3",
-    projectProviders = HardenXStreamCodemodTestProjectProvider.class)
-final class HardenXStreamCodemodTestDenyTypes implements CodemodTestMixin {}
+    projectProviders = HardenXStreamCodemodTestProjectProvider.XStreamSameDependencyVersion.class)
+final class HardenXStreamCodemodTestDenyTypesSameDependencyVersion implements CodemodTestMixin {}
+
+@Metadata(
+    codemodType = HardenXStreamCodemod.class,
+    testResourceDir = "harden-xstream-denyTypes",
+    dependencies = "io.github.pixee:java-security-toolkit:1.1.3",
+    projectProviders =
+        HardenXStreamCodemodTestProjectProvider.XStreamGreaterDependencyVersion.class)
+final class HardenXStreamCodemodTestDenyTypesGreaterDependencyVersion implements CodemodTestMixin {}
+
+@Metadata(
+    codemodType = HardenXStreamCodemod.class,
+    testResourceDir = "harden-xstream",
+    dependencies = "io.github.pixee:java-security-toolkit-xstream:1.0.2",
+    projectProviders = HardenXStreamCodemodTestProjectProvider.XStreamLowerDependencyVersion.class)
+final class HardenXStreamCodemodTestDenyTypesLowerDependencyVersion implements CodemodTestMixin {}
